test_projects/openai_gpt_4o_mini_python/requirements.txt

Flask==2.0.3
psycopg2-binary==2.9.1
Werkzeug==2.0.3
